# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MT-Engine",
    "description": "M-Team \u514d\u8d39\u79cd\u5b50\u730e\u624b",
    "version": "6.6.2"
  },
  "paths": {
...
    "/api/panel/torrents": {
      "get": {
        "summary": "Get Panel Torrents",
        "description": "\u83b7\u53d6\u79cd\u5b50\u5217\u8868\uff08\u5e26\u7b5b\u9009\uff09\n\nArgs:\n    tag: \u6807\u7b7e\u7b5b\u9009 (\"\u58f0\u5450\u505a\u79cd\" | \"\u96f7\u8fbe\u4e0b\u8f7d\" | \"PILOT\")\n    status: \u72b6\u6001\u7b5b\u9009 (\"downloading\" | \"seeding\" | \"paused\" | \"completed\")\n\nReturns:\n    {\n        \"torrents\": List[Dict],\n        \"total_count\": int,\n        \"filtered_count\": int\n    }",
        "operationId": "get_panel_torrents_api_panel_torrents_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelTorrentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/torrents": {
      "get": {
        "summary": "Get Torrents",
        "description": "\u83b7\u53d6\u79cd\u5b50\u5217\u8868",
        "operationId": "get_torrents_api_torrents_get",
        "parameters": [
          {
            "name": "discount",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Discount"
            }
          },
          {
            "name": "min_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Size"
            }
          },
          {
            "name": "max_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Size"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/refresh": {
      "post": {
        "summary": "Refresh",
        "description": "\u624b\u52a8\u89e6\u53d1\u5237\u65b0",
        "operationId": "refresh_api_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/download": {
      "post": {
        "summary": "Download Torrent",
        "description": "\u4ece Free Hunter \u4e0b\u8f7d\u79cd\u5b50",
        "operationId": "download_torrent_api_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/auto-delete/toggle": {
      "post": {
        "summary": "Auto Delete Toggle",
        "description": "\u5207\u6362\u81ea\u52a8\u5220\u9664\u529f\u80fd",
        "operationId": "auto_delete_toggle_api_auto_delete_toggle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoDeleteToggleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/radar": {
      "post": {
        "summary": "Radar",
        "description": "\u96f7\u8fbe\u641c\u7d22\u79cd\u5b50",
        "operationId": "radar_api_radar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/radar/download": {
      "post": {
        "summary": "Radar Download",
        "description": "\u4ece\u96f7\u8fbe\u7ed3\u679c\u4e0b\u8f7d\u79cd\u5b50",
        "operationId": "radar_download_api_radar_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoDeleteToggleRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "AutoDeleteToggleRequest",
        "description": "Request model for auto-delete toggle (idempotent set)"
      },
...
      "DownloadRequest": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^\\d+$",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DownloadRequest",
        "description": "Request model for torrent download"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PanelTorrentsResponse": {
        "properties": {
          "torrents": {
            "items": {
              "$ref": "#/components/schemas/PanelTorrentItem"
            },
            "type": "array",
            "title": "Torrents"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "default": 0
          },
          "filtered_count": {
            "type": "integer",
            "title": "Filtered Count",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "PanelTorrentsResponse",
        "description": "PANEL torrent list response"
      },
...
      "SearchRequest": {
        "properties": {
          "keyword": {
            "type": "string",
            "maxLength": 100,
            "title": "Keyword",
            "default": ""
          },
          "mode": {
            "type": "string",
            "enum": [
              "normal",
              "adult",
              "movie",
              "tvshow",
              "other"
            ],
            "title": "Mode",
            "default": "normal"
          },
          "categories": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Categories"
          },
          "standards": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Standards"
          },
          "videoCodecs": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Videocodecs"
          },
          "audioCodecs": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Audiocodecs"
          },
          "sources": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Sources"
          },
          "countries": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Countries"
          },
          "discount": {
            "type": "string",
            "title": "Discount",
            "default": ""
          },
          "sortField": {
            "type": "string",
            "title": "Sortfield",
            "default": "CREATED_DATE"
          },
          "sortDirection": {
            "type": "string",
            "title": "Sortdirection",
            "default": "DESC"
          },
          "pageNumber": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Pagenumber",
            "default": 1
          },
          "pageSize": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Pagesize",
            "default": 50
          }
        },
        "type": "object",
        "title": "SearchRequest",
        "description": "Request model for torrent search"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "MT Engine API key sent as an Authorization bearer token.",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "MT Engine API key sent in the X-MT-ENGINE-Key header.",
        "in": "header",
        "name": "X-MT-ENGINE-Key"
      }
    }
  }
}